refactor(macros,core,backends): the macro writes the attribute accessors
`operation!` declared an attribute's name and kind but emitted nothing to
read or write it, so ops hand-wrote wrappers around `.attr("name")` and
`.attr("name", AttributeValue::...)`. Each declared attribute of a scalar
kind now earns a typed setter on the builder and a typed getter on the
operation, both named after it; `Array`, `Dict` and the untyped kinds keep
going through `attr`, having no single Rust type.

The `Predicate`-only setter the macro used to emit is subsumed by the
general one. Twenty-seven hand-written wrappers and their private helpers
are gone: `SymbolAddressOp::sym_name`, four `outgoing_stack_size`, the `cfg`
branch destinations and their `block_attr`, `block_attr`/`string_attr` in
`target_boilerplate`, `attr_str`/`attr_type` in the SPIR-V writer,
`string_attribute`/`uint_attribute` in the `cir` lowering, and the
alloca, tuple, global, func and constant accessors that the generated ones
now name.

`dialect!` spells `<Op as Operation>::name()` so an op carrying a `name`
attribute keeps its parser registration.

+++ core/src/passes/inline.rs
@@ -479,7 +479,7 @@ fn fold_tuple_gets(
         let Some(made) = context.get_op(source).as_op::<MakeTupleOp>() else {
             continue;
         };
-        let Some(&element) = Operation::operands(&made).get(get.index()) else {
+        let Some(&element) = Operation::operands(&made).get(get.index() as usize) else {
             continue;
         };
         rename(context, body, get.result(), element);
